# contracts/generated/openapi/contracts.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RiskItem ModelOps API",
    "version": "0.1.0",
    "description": "信贷风控模型智能监测与自主迭代系统 API。由 scripts/generate_openapi.py 从 FastAPI app 自动生成，不手工维护。"
  },
  "paths": {
...
    "/api/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_models_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(ACTIVE|INACTIVE|RETIRED)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Register Model",
        "operationId": "register_model_api_models_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "operationId": "get_model_api_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/versions": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Register Version",
        "operationId": "register_version_api_models__model_id__versions_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/data-windows": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Data Windows",
        "operationId": "list_data_windows_api_models__model_id__data_windows_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lifecycle-runs": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Start Run",
        "operationId": "start_run_api_lifecycle_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lifecycle-runs/{lifecycle_run_id}": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_lifecycle_runs__lifecycle_run_id__get",
        "parameters": [
          {
            "name": "lifecycle_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lifecycle Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lifecycle-runs/{lifecycle_run_id}/resume": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Resume Run",
        "operationId": "resume_run_api_lifecycle_runs__lifecycle_run_id__resume_post",
        "parameters": [
          {
            "name": "lifecycle_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lifecycle Run Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeRequest",
                "default": {
                  "decision": "approved"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lifecycle-runs/{lifecycle_run_id}/cancel": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Cancel Run",
        "operationId": "cancel_run_api_lifecycle_runs__lifecycle_run_id__cancel_post",
        "parameters": [
          {
            "name": "lifecycle_run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lifecycle Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/resolve-alert": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Resolve Alert",
        "description": "查询 Metric → Alert 映射。",
        "operationId": "resolve_alert_api_knowledge_resolve_alert_get",
        "parameters": [
          {
            "name": "metric_code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "指标代码，如 FEATURE_PSI",
              "title": "Metric Code"
            },
            "description": "指标代码，如 FEATURE_PSI"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/entities/{entity_code}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Entity",
        "description": "查询单个知识实体。",
        "operationId": "get_entity_api_knowledge_entities__entity_code__get",
        "parameters": [
          {
            "name": "entity_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegisterModelRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Id"
          },
          "model_name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Model Name"
          },
          "model_type": {
            "type": "string",
            "title": "Model Type",
            "default": "CREDIT_RISK"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "model_name"
        ],
        "title": "RegisterModelRequest",
        "description": "POST /api/models 请求体。"
      },
      "RegisterVersionRequest": {
        "properties": {
          "version_code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Version Code"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "CHALLENGER"
          }
        },
        "type": "object",
        "required": [
          "version_code"
        ],
        "title": "RegisterVersionRequest",
        "description": "POST /api/models/{model_id}/versions 请求体。"
      },
...
      "ResumeRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision",
            "default": "approved"
          },
          "resume_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume Type"
          },
          "manual_review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Review Id"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          },
          "plan_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Type"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "repair_plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repair Plan Id"
          },
          "artifact_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Uri"
          },
          "metrics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metrics"
          },
          "best_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Threshold"
          },
          "training_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Job Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "worker_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Status"
          },
          "candidate_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Version"
          },
          "experiment_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment Id"
          }
        },
        "type": "object",
        "title": "ResumeRequest"
      },
...
      "StartRunRequest": {
        "properties": {
          "model_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Id"
          },
          "champion_version": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Champion Version"
          },
          "trigger_type": {
            "type": "string",
            "title": "Trigger Type",
            "default": "SCHEDULED_TRIGGER"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "champion_version"
        ],
        "title": "StartRunRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}